Updated teams

+++ API/src/main/java/org/setup/Listify/service/TeamsService.java
@@ -6,14 +6,16 @@ import org.setup.Listify.exception.*;
 import org.setup.Listify.model.TeamMembers;
 import org.setup.Listify.dto.UserTeamProjects;
 import org.setup.Listify.model.Teams;
+import org.setup.Listify.model.Users;
 import org.setup.Listify.repo.TeamMembersRepository;
 import org.setup.Listify.repo.TeamsRepository;
+import org.setup.Listify.repo.UsersRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TeamsService {
@@ -24,86 +26,77 @@ public class TeamsService {
     @Autowired
     private TeamMembersRepository teamMembersRepository;
     @Autowired
+    private UsersRepository usersRepository;
+    @Autowired
     UserService userService;
 
     public List<Teams> getAllUserTeams(Long userID) {
+        if (userID == null) {
+            throw new UnauthorizedException("User is not authenticated. Please log in.");
+        }
+
         List<Teams> teams = teamsRepository.findTeamsByUserID(userID);
 
         if (teams.isEmpty()) {
-            throw new ListNotFoundException("teams");
+            throw new NotFoundException("No teams for user " + userID + " were found");
         }
         return teams;
     }
 
-    public List<TeamMembers> getAllTeamMembers() {
-        List<TeamMembers> teamMembers = teamMembersRepository.findAll();
-
-        if (teamMembers.isEmpty()) {
-            throw new ListNotFoundException("team members");
+    public void checkTeamExistence(Long teamID) {
+        if (teamsRepository.findByTeamID(teamID) == null) {
+            throw new NotFoundException("Team with id " + teamID + " was not found");
         }
-
-        return teamMembers;
     }
 
-    public Teams getTeamById(Long teamID) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        Long userID = userService.getUserIDFromAuthentication(authentication);
 
-        Teams team = teamsRepository.getTeamById(teamID, userID);
+    public Teams findATeamByUserID(Long userID, Long teamID) {
+        if (userID == null) {
+            throw new UnauthorizedException("User is not authenticated. Please log in.");
+        }
+
+        checkTeamExistence(teamID);
+
+        Teams team = teamsRepository.findATeamByUserID(userID, teamID);
 
         if (team == null) {
-            throw new TeamNotFoundException(teamID);
+            throw new ForbiddenException("User " + userID + " is not added to the team " + teamID);
         }
 
         return team;
     }
 
-    public TeamMembers getByTeamMembersId(Long userID) {
-        return teamMembersRepository.findById(userID).orElseThrow(() -> new TeamMemberNotFoundException(userID));
-    }
-
-    public TeamMembers getTeamMemberByUserId(Long userID) {
-        List<TeamMembers> teamMembersList = teamMembersRepository.findByUserID(userID);
-
-        if (teamMembersList.isEmpty()) {
-            throw new UserNotFoundException("User with id: " + userID + " not found");
-        }
-        return teamMembersList.getFirst();
-    }
-
-    public List<TeamMembers> getTeamMembersByTeamId(Long teamID) {
-        getTeamById(teamID);
+    public List<TeamMembers> getTeamMembersByTeamID(Long teamID, Long userID) {
+        findATeamByUserID(userID, teamID);
         List<TeamMembers> teamMembers = teamMembersRepository.findByTeamID(teamID);
 
         if (teamMembers.isEmpty()) {
-            throw new ListNotFoundException("team members for team " + teamID);
+            throw new NotFoundException("No members for this team was found " + teamID);
         }
 
         return teamMembers;
     }
 
     public boolean isTeamLeader(Long teamID, Long userID) {
-        boolean isTeamLeader = teamMembersRepository.existsByTeamIDAndUserIDAndIsTeamLeaderTrue(teamID, userID);
-        if (!isTeamLeader) {
-            throw new NotTeamLeaderException("User with ID " + userID + " is not the team leader for team ID " + teamID);
-        }
-        return isTeamLeader;
+        return teamMembersRepository.existsByTeamIDAndUserIDAndIsTeamLeaderTrue(teamID, userID);
     }
 
-
     public List<UserTeamProjects> getProjectsByTeamIDAndUserID(Long teamID, Long userID) {
-        getTeamById(teamID);
+        findATeamByUserID(userID, teamID);
         List<UserTeamProjects> userTeamProjects = teamsRepository.findProjectsByTeamIDAndUserID(teamID, userID);
 
         if (userTeamProjects.isEmpty()) {
-            throw new ListNotFoundException("team projects for team " + teamID);
+            throw new NotFoundException("No project found for team " + teamID);
         }
 
         return userTeamProjects;
     }
 
+    @Transactional
     public Long addTeam(Long userID, String teamName) {
-        TeamMembers teamMember = getTeamMemberByUserId(userID);
+        if (userID == null) {
+            throw new UnauthorizedException("User is not authenticated. Please log in");
+        }
 
         teamsRepository.createTeam(userID, teamName);
 
@@ -111,56 +104,101 @@ public class TeamsService {
         return lastCreatedTeam != null ? lastCreatedTeam.getTeamID() : null;
     }
 
-    public void assignMemberToTeam(Long teamLeaderID, Long userID, Long teamID) {
-        getTeamById(teamID);
-        getTeamMemberByUserId(userID);
-        isTeamLeader(teamID, teamLeaderID);
-        boolean memberExists = teamMembersRepository.existsByTeamIDAndUserID(teamID, userID);
+    @Transactional
+    public void assignMemberToTeam(Long teamLeaderID, String githubID, Long teamID) {
+        Long userID = getUserIDFromGithubID(githubID);
 
+        Teams team = findATeamByUserID(teamLeaderID, teamID);
+
+        if (team.getCreatedAt().equals(userService.getUserCreatedAt(teamLeaderID))) {
+            throw new ForbiddenException("User can't assign members to personal team");
+        }
+
+        boolean isTeamLeader = isTeamLeader(teamID, teamLeaderID);
+        if (!isTeamLeader) {
+            throw new ForbiddenException("Only a team leader can add new members to a team");
+        }
+
+        boolean memberExists = teamMembersRepository.existsByTeamIDAndUserID(teamID, userID);
         if (memberExists) {
-            throw new TeamMemberAlreadyExistsException("User with ID " + userID + " is already a member of team " + teamID);
+            throw new ForbiddenException("User with ID " + userID + " is already a member of team " + teamID);
         }
 
         teamMembersRepository.assignMemberToTeam(teamLeaderID, userID, teamID);
     }
 
+    @Transactional
     public void deleteTeam(Long teamID, Long teamLeaderID) {
-        getTeamById(teamID);
-        isTeamLeader(teamID, teamLeaderID);
+        Teams team = findATeamByUserID(teamLeaderID, teamID);
+        boolean isTeamLeader = isTeamLeader(teamID, teamLeaderID);
+
+        if (!isTeamLeader) {
+            throw new ForbiddenException("Only the team leader can delete a team");
+        }
+
+        if (team.getCreatedAt().equals(userService.getUserCreatedAt(teamLeaderID))) {
+            throw new ForbiddenException("User can't delete personal team");
+        }
+
         teamsRepository.deleteTeam(teamID, teamLeaderID);
     }
 
-    public void deleteMemberFromTeam(Long userID, Long teamID, Long teamLeaderID) {
-        getTeamById(teamID);
-        isTeamLeader(teamID, teamLeaderID);
-        boolean memberExists = teamMembersRepository.existsByTeamIDAndUserID(teamID, userID);
-        if (userID == teamID) {
-            throw new DeleteTeamLeaderException();
+    @Transactional
+    public void deleteMemberFromTeam(String githubID, Long teamID, Long teamLeaderID) {
+        Long userID = getUserIDFromGithubID(githubID);
+
+        boolean isTeamLeader = isTeamLeader(teamID, teamLeaderID);
+        if (!isTeamLeader) {
+            throw new ForbiddenException("Only team leaders can delete a member from a team");
         }
 
-        if (!memberExists) {
-            throw new TeamMemberAlreadyExistsException("User with ID " + userID + " is not a member of team " + teamID);
+        if (userID.equals(teamLeaderID)) {
+            throw new ForbiddenException("A team leader can't be deleted from the team, please first re-assign team leader");
         }
+
+        boolean memberExists = teamMembersRepository.existsByTeamIDAndUserID(teamID, userID);
+        if (!memberExists) {
+            throw new ForbiddenException("User with ID " + userID + " is not a member of team " + teamID);
+        }
+
         teamMembersRepository.deleteMemberFromTeam(userID, teamID, teamLeaderID);
     }
 
     @Transactional
-    public void updateTeamDetails(Long userID, Long teamID, String newTeamName) {
-        getTeamById(teamID);
-        isTeamLeader(teamID, userID);
+    public Teams updateTeamDetails(Long userID, Long teamID, String newTeamName) {
+        findATeamByUserID(userID, teamID);
+        boolean isTeamLeader = isTeamLeader(teamID, userID);
+
+        if (!isTeamLeader) {
+            throw new ForbiddenException("Only team leaders can update teams details");
+        }
+
         teamsRepository.updateTeamDetails(userID, teamID, newTeamName);
         entityManager.flush();
         entityManager.clear();
+
+        return findATeamByUserID(userID, teamID);
     }
 
-    public void updateTeamLeader(Long teamLeaderID, Long teamID, Long newTeamLeaderID) {
-        getTeamById(teamID);
-        isTeamLeader(teamID, teamLeaderID);
-        boolean memberExists = teamMembersRepository.existsByTeamIDAndUserID(teamID, newTeamLeaderID);
+    public void updateTeamLeader(Long teamLeaderID, Long teamID, String newTeamLeaderGithubID) {
+        Long newTeamLeaderID = getUserIDFromGithubID(newTeamLeaderGithubID);
 
-        if (!memberExists) {
-            throw new TeamMemberAlreadyExistsException("User with ID " + newTeamLeaderID + " is not a member of team " + teamID);
+        findATeamByUserID(newTeamLeaderID, teamID);
+        boolean isTeamLeader = isTeamLeader(teamID, teamLeaderID);
+
+        if (!isTeamLeader) {
+            throw new ForbiddenException("Only a team leader can reassign a team leader");
         }
+
         teamMembersRepository.updateTeamLeader(teamLeaderID, teamID, newTeamLeaderID);
+    }
+
+    public Long getUserIDFromGithubID(String githubID) {
+        Optional<Users> users = usersRepository.findByGitHubID(githubID);
+        if (users.isEmpty()) {
+            throw new NotFoundException("User with github ID: " + githubID + " not found");
+        }
+
+        return users.get().getUserID();
     }
 }
